further move views related to events from group to the event view
