+++ server/apps/event/views.py
@@ -1,11 +1,14 @@
-from django.shortcuts import redirect
+from datetime import date
+
+from django.shortcuts import redirect, render
+from django.contrib import messages
 from django.contrib.auth.models import User
-from django.views.generic.base import TemplateView
+from django.views.generic.base import TemplateView, View
 from django.views.generic import UpdateView, FormView
 from django.contrib.auth.decorators import login_required
 
 from .models import *
-from .forms import EventForm
+from .forms import EventForm, EventFormSet
 
 from apps.group.models import Group
 from apps.utils.accessMixins import LoginRequiredAccessMixin, UserIsAdmin
@@ -69,6 +72,42 @@ class EventUpdateView(UserIsAdmin, UpdateView):
         return super().dispatch(request, *args, **kwargs)
 
 
+class UpdateGroupEventsView(UserIsAdmin, View):
+    template_name = 'group/event/planned_edit.html'
+
+    def get_context_data(self, **kwargs):
+        context = {}
+        context['object'] = Group.get_group_by_slug(kwargs['group_slug'])
+        context['events'] = BaseEvent.objects.filter(
+            group=kwargs['group_slug'], date__gte=date.today())
+        context['form'] = EventFormSet(queryset=context['events'])
+        return context
+
+    def get(self, request, group_slug):
+        return render(request, self.template_name, context=self.get_context_data(group_slug=group_slug))
+
+    def post(self, request,  group_slug):
+        return edit_events(request, group_slug)
+
+
+class UpdateGroupArchivedEventsView(UserIsAdmin, View):
+    template_name = 'group/archived_edit.html'
+
+    def get_context_data(self, **kwargs):
+        context = {}
+        context['object'] = Group.get_group_by_slug(kwargs['group_slug'])
+        context['events'] = BaseEvent.objects.filter(
+            group=kwargs['group_slug'], date__lt=date.today())
+        context['form'] = EventFormSet(queryset=context['events'])
+        return context
+
+    def get(self, request, group_slug):
+        return render(request, self.template_name, context=self.get_context_data(group_slug=group_slug))
+
+    def post(self, request,  group_slug):
+        return edit_events(request, group_slug)
+
+
 @login_required
 def add_participant(request, event_slug):
     """Adds the user to the list of participants."""
@@ -87,3 +126,23 @@ def remove_participant(request, event_slug):
     if request.GET.get('redirect'):
         return redirect('home:home')
     return redirect(event.get_absolute_url())
+
+
+@login_required
+def edit_events(request, group_slug):
+    group = Group.get_group_by_slug(group_slug)
+    form = EventFormSet(request.POST)
+    if form.is_valid():
+        events = form.save(commit=False)
+        # Link each event to the group
+        for event in events:
+            event.group = group.slug
+            event.save()
+        # Delete  missing events
+        for event in form.deleted_objects:
+            event.delete()
+        messages.success(request, 'Events  modifies')
+        return redirect('group:update-events', group_slug)
+    else:
+        messages.warning(request, form.errors)
+        return redirect('group:update-events', group_slug)
